Rename asset_condition folder and add README
