+++ python_modules/dagster-graphql/dagster_graphql/schema/asset_condition_evaluations.py
@@ -3,8 +3,10 @@ import itertools
 from typing import Optional, Sequence, Union
 
 import graphene
-from dagster._core.definitions.asset_condition.asset_condition import AssetConditionEvaluation
 from dagster._core.definitions.asset_subset import AssetSubset
+from dagster._core.definitions.declarative_scheduling.asset_condition import (
+    AssetConditionEvaluation,
+)
 from dagster._core.definitions.partition import PartitionsDefinition, PartitionsSubset
 from dagster._core.definitions.time_window_partitions import BaseTimeWindowPartitionsSubset
 from dagster._core.scheduler.instigation import AutoMaterializeAssetEvaluationRecord
