+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_evaluation.py
@@ -34,7 +34,7 @@ from dagster._utils.security import non_secure_md5_hash_str
 from .partition import PartitionsDefinition, SerializedPartitionsSubset
 
 if TYPE_CHECKING:
-    from .asset_condition.asset_condition import (
+    from .declarative_scheduling.asset_condition import (
         AssetConditionEvaluation,
         AssetConditionEvaluationWithRunIds,
         AssetConditionSnapshot,
@@ -148,7 +148,7 @@ def deserialize_auto_materialize_asset_evaluation_to_asset_condition_evaluation_
     """Provides a backcompat layer to allow deserializing old AutoMaterializeAssetEvaluation
     objects into the new AssetConditionEvaluationWithRunIds objects.
     """
-    from .asset_condition.asset_condition import AssetConditionEvaluationWithRunIds
+    from .declarative_scheduling.asset_condition import AssetConditionEvaluationWithRunIds
 
     class BackcompatDeserializer(BackcompatAutoMaterializeAssetEvaluationSerializer):
         @property
@@ -213,7 +213,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
     def _asset_condition_snapshot_from_rule_snapshot(
         self, rule_snapshot: AutoMaterializeRuleSnapshot
     ) -> "AssetConditionSnapshot":
-        from .asset_condition.asset_condition import AssetConditionSnapshot, RuleCondition
+        from .declarative_scheduling.asset_condition import AssetConditionSnapshot, RuleCondition
 
         unique_id_parts = [rule_snapshot.class_name, rule_snapshot.description]
         unique_id = non_secure_md5_hash_str("".join(unique_id_parts).encode())
@@ -233,7 +233,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         is_partitioned: bool,
         rule_snapshot: AutoMaterializeRuleSnapshot,
     ) -> "AssetConditionEvaluation":
-        from .asset_condition.asset_condition import (
+        from .declarative_scheduling.asset_condition import (
             AssetConditionEvaluation,
             AssetSubsetWithMetadata,
             HistoricalAllPartitionsSubsetSentinel,
@@ -278,7 +278,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         is_partitioned: bool,
         decision_type: AutoMaterializeDecisionType,
     ) -> Optional["AssetConditionEvaluation"]:
-        from .asset_condition.asset_condition import (
+        from .declarative_scheduling.asset_condition import (
             AssetConditionEvaluation,
             AssetConditionSnapshot,
             HistoricalAllPartitionsSubsetSentinel,
@@ -371,7 +371,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         whitelist_map: WhitelistMap,
         context: UnpackContext,
     ) -> "AssetConditionEvaluationWithRunIds":
-        from .asset_condition.asset_condition import (
+        from .declarative_scheduling.asset_condition import (
             AndAssetCondition,
             AssetConditionEvaluation,
             AssetConditionSnapshot,
