+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_impls.py
@@ -47,11 +47,11 @@ from dagster._utils.schedules import (
 from .base_asset_graph import sort_key_for_asset_partition
 
 if TYPE_CHECKING:
-    from dagster._core.definitions.asset_condition.asset_condition import (
+    from dagster._core.definitions.declarative_scheduling.asset_condition import (
         AssetConditionResult,
     )
 
-    from .asset_condition.scheduling_condition_evaluation_context import (
+    from .declarative_scheduling.scheduling_condition_evaluation_context import (
         SchedulingConditionEvaluationContext,
     )
 
@@ -77,7 +77,7 @@ class MaterializeOnRequiredForFreshnessRule(
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         true_subset, subsets_with_metadata = freshness_evaluation_results_for_asset_key(
             context.legacy_context.root_context
@@ -195,7 +195,7 @@ class MaterializeOnCronRule(
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         missed_ticks = self.missed_cron_ticks(context)
         new_asset_partitions = self.get_new_candidate_asset_partitions(context, missed_ticks)
@@ -353,7 +353,7 @@ class MaterializeOnParentUpdatedRule(
         """Evaluates the set of asset partitions of this asset whose parents have been updated,
         or will update on this tick.
         """
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         asset_partitions_by_updated_parents: Mapping[
             AssetKeyPartitionKey, Set[AssetKeyPartitionKey]
@@ -480,7 +480,7 @@ class MaterializeOnMissingRule(AutoMaterializeRule, NamedTuple("_MaterializeOnMi
         """Evaluates the set of asset partitions for this asset which are missing and were not
         previously discarded.
         """
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         if (
             context.legacy_context.asset_key
@@ -561,7 +561,7 @@ class SkipOnParentOutdatedRule(AutoMaterializeRule, NamedTuple("_SkipOnParentOut
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -614,7 +614,7 @@ class SkipOnParentMissingRule(AutoMaterializeRule, NamedTuple("_SkipOnParentMiss
         self,
         context: "SchedulingConditionEvaluationContext",
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -690,7 +690,7 @@ class SkipOnNotAllParentsUpdatedRule(
         self,
         context: "SchedulingConditionEvaluationContext",
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -917,7 +917,7 @@ class SkipOnNotAllParentsUpdatedSinceCronRule(
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         passed_time_window = self.passed_time_window(context)
         has_new_passed_time_window = passed_time_window.end.timestamp() > (
@@ -993,7 +993,7 @@ class SkipOnRequiredButNonexistentParentsRule(
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         asset_partitions_by_evaluation_data = defaultdict(set)
 
@@ -1046,7 +1046,7 @@ class SkipOnBackfillInProgressRule(
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         backfilling_subset = (
             # this backfilling subset is aware of the current partitions definitions, and so will
@@ -1083,7 +1083,7 @@ class DiscardOnMaxMaterializationsExceededRule(
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         # the set of asset partitions which exceed the limit
         rate_limited_asset_partitions = set(
@@ -1116,7 +1116,7 @@ class SkipOnRunInProgressRule(AutoMaterializeRule, NamedTuple("_SkipOnRunInProgr
     def evaluate_for_asset(
         self, context: "SchedulingConditionEvaluationContext"
     ) -> "AssetConditionResult":
-        from .asset_condition.asset_condition import AssetConditionResult
+        from .declarative_scheduling.asset_condition import AssetConditionResult
 
         if context.legacy_context.partitions_def is not None:
             raise DagsterInvariantViolationError(
