+++ python_modules/dagster/dagster/_core/definitions/declarative_scheduling/asset_condition_evaluation_context.py
@@ -19,7 +19,7 @@ from typing import (
 
 import pendulum
 
-from dagster._core.definitions.asset_condition.asset_condition import (
+from dagster._core.definitions.declarative_scheduling.asset_condition import (
     HistoricalAllPartitionsSubsetSentinel,
 )
 from dagster._core.definitions.events import AssetKey, AssetKeyPartitionKey
