+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/test_auto_materialize_asset_evaluation.py
@@ -1,9 +1,9 @@
 from dagster import DailyPartitionsDefinition, MetadataValue
-from dagster._core.definitions.asset_condition.asset_condition import AssetSubsetWithMetadata
 from dagster._core.definitions.asset_subset import AssetSubset
 from dagster._core.definitions.auto_materialize_rule_evaluation import (
     deserialize_auto_materialize_asset_evaluation_to_asset_condition_evaluation_with_run_ids,
 )
+from dagster._core.definitions.declarative_scheduling.asset_condition import AssetSubsetWithMetadata
 from dagster._core.definitions.events import AssetKey
