+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/custom_condition_scenarios.py
@@ -1,6 +1,9 @@
 from dagster import AutoMaterializeRule
-from dagster._core.definitions.asset_condition.asset_condition import AssetCondition, RuleCondition
 from dagster._core.definitions.auto_materialize_policy import AutoMaterializePolicy
+from dagster._core.definitions.declarative_scheduling.asset_condition import (
+    AssetCondition,
+    RuleCondition,
+)
 
 from ..base_scenario import run_request
 from ..scenario_specs import one_asset, two_assets_in_sequence
